requirements.txt
tqdm==4.51.0
gensim==3.8.2
streamlit==0.69.2
Unidecode==1.1.1
Werkzeug==1.0.1
Flask-WTF==0.14.3
numpy==1.18.5
pyttsx3==2.90
Flask==1.1.2
nltk==3.5
opencv-python==4.4.0.42
pytesseract==0.3.6
dataclasses==0.6
transformers==3.0.0
apex==0.9.10dev
beautifulsoup4==4.9.3
nlp==0.4
Pillow==8.0.1
scikit-learn==0.23.2
gunicorn==19.6.0
torch==1.6.0